+++ common/src/main/java/de/cech12/unlitcampfire/mixinaccess/ICampfireBlockMixin.java
@@ -8,11 +8,11 @@ public interface ICampfireBlockMixin {
     BooleanProperty INFINITE = BooleanProperty.create("infinite");
     BooleanProperty RUNS_OUT = BooleanProperty.create("runs_out");
 
-    int unlitCampfire$getMaxLitTime(BlockState state);
+    long unlitCampfire$getMaxLitTime(BlockState state);
 
-    int unlitCampfire$getMaxLitTimeExtension(BlockState state);
+    long unlitCampfire$getMaxLitTimeExtension(BlockState state);
 
-    int unlitCampfire$getRunsOutIndicatorTime(BlockState state);
+    long unlitCampfire$getRunsOutIndicatorTime(BlockState state);
 
     boolean unlitCampfire$burnsInfinite(BlockState state);
