+++ fabric/src/main/java/de/cech12/unlitcampfire/platform/FabricConfigHelper.java
@@ -21,13 +21,13 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     public static final class Campfire implements ConfigData {
 
         @ConfigEntry.Gui.Tooltip(count = 6)
-        public int CAMPFIRE_LIT_TIME = CAMPFIRE_LIT_TIME_DEFAULT;
+        public long CAMPFIRE_LIT_TIME = CAMPFIRE_LIT_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 7)
-        public int CAMPFIRE_RUN_OUT_INDICATOR_TIME = CAMPFIRE_RUN_OUT_INDICATOR_TIME_DEFAULT;
+        public long CAMPFIRE_RUN_OUT_INDICATOR_TIME = CAMPFIRE_RUN_OUT_INDICATOR_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 7)
-        public int CAMPFIRE_RAIN_UNLIT_TIME = CAMPFIRE_RAIN_UNLIT_TIME_DEFAULT;
+        public long CAMPFIRE_RAIN_UNLIT_TIME = CAMPFIRE_RAIN_UNLIT_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 6)
         public int CAMPFIRE_RAIN_PARTICLE_FACTOR = CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT;
@@ -39,7 +39,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
         public boolean CAMPFIRE_ADDING_BURNABLES = CAMPFIRE_ADDING_BURNABLES_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 5)
-        public int CAMPFIRE_MAX_LIT_TIME = CAMPFIRE_MAX_LIT_TIME_EXTENSION_DEFAULT;
+        public long CAMPFIRE_MAX_LIT_TIME = CAMPFIRE_MAX_LIT_TIME_EXTENSION_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 4)
         public boolean CAMPFIRE_AFFECTED_BY_SLEEP_TIME = CAMPFIRE_AFFECTED_BY_SLEEP_TIME_DEFAULT;
@@ -60,13 +60,13 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     public static final class SoulCampfire implements ConfigData {
 
         @ConfigEntry.Gui.Tooltip(count = 6)
-        public int SOUL_CAMPFIRE_LIT_TIME = SOUL_CAMPFIRE_LIT_TIME_DEFAULT;
+        public long SOUL_CAMPFIRE_LIT_TIME = SOUL_CAMPFIRE_LIT_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 7)
-        public int SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME = SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME_DEFAULT;
+        public long SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME = SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 7)
-        public int SOUL_CAMPFIRE_RAIN_UNLIT_TIME = SOUL_CAMPFIRE_RAIN_UNLIT_TIME_DEFAULT;
+        public long SOUL_CAMPFIRE_RAIN_UNLIT_TIME = SOUL_CAMPFIRE_RAIN_UNLIT_TIME_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 6)
         public int SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR = SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT;
@@ -78,7 +78,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
         public boolean SOUL_CAMPFIRE_ADDING_BURNABLES = SOUL_CAMPFIRE_ADDING_BURNABLES_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 5)
-        public int SOUL_CAMPFIRE_MAX_LIT_TIME = SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION_DEFAULT;
+        public long SOUL_CAMPFIRE_MAX_LIT_TIME = SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION_DEFAULT;
 
         @ConfigEntry.Gui.Tooltip(count = 4)
         public boolean SOUL_CAMPFIRE_AFFECTED_BY_SLEEP_TIME = SOUL_CAMPFIRE_AFFECTED_BY_SLEEP_TIME_DEFAULT;
@@ -102,7 +102,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     }
 
     @Override
-    public int getLitTime(boolean isSoulCampfire) {
+    public long getLitTime(boolean isSoulCampfire) {
         FabricConfigHelper config = getConfig();
         return isSoulCampfire ?
                 Math.clamp(config.SOUL_CAMPFIRE.SOUL_CAMPFIRE_LIT_TIME, SOUL_CAMPFIRE_LIT_TIME_MIN, SOUL_CAMPFIRE_LIT_TIME_MAX) :
@@ -110,7 +110,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     }
 
     @Override
-    public int getRunOutIndicatorTime(boolean isSoulCampfire) {
+    public long getRunOutIndicatorTime(boolean isSoulCampfire) {
         FabricConfigHelper config = getConfig();
         return isSoulCampfire ?
                 Math.clamp(config.SOUL_CAMPFIRE.SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME, SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME_MIN, SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME_MAX) :
@@ -118,7 +118,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     }
 
     @Override
-    public int getRainUnlitTime(boolean isSoulCampfire) {
+    public long getRainUnlitTime(boolean isSoulCampfire) {
         FabricConfigHelper config = getConfig();
         return isSoulCampfire ?
                 Math.clamp(config.SOUL_CAMPFIRE.SOUL_CAMPFIRE_RAIN_UNLIT_TIME, SOUL_CAMPFIRE_RAIN_UNLIT_TIME_MIN, SOUL_CAMPFIRE_RAIN_UNLIT_TIME_MAX) :
@@ -146,7 +146,7 @@ public class FabricConfigHelper implements ConfigData, IConfigHelper {
     }
 
     @Override
-    public int getMaxLitTimeExtension(boolean isSoulCampfire) {
+    public long getMaxLitTimeExtension(boolean isSoulCampfire) {
         FabricConfigHelper config = getConfig();
         return isSoulCampfire ?
                 Math.clamp(config.SOUL_CAMPFIRE.SOUL_CAMPFIRE_MAX_LIT_TIME, SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION_MIN, SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION_MAX) :
